feat(sdk-experiments): fix comments
