+++ core-web/libs/sdk/experiments/.eslintrc.json
@@ -25,13 +25,6 @@
                     }
                 ]
             }
-        },
-        {
-            "files": ["*.spec.ts", "*.test.ts"],
-            "rules": {
-                "@typescript-eslint/no-explicit-any": "off",
-                "@typescript-eslint/ban-ts-comment": "off"
-            }
         }
     ]
 }
