+++ core-web/nx.json
@@ -107,6 +107,11 @@
             "inputs": ["default", "{workspaceRoot}/.eslintrc.json"],
             "cache": true
         },
+        "@nx/rollup:rollup": {
+            "cache": true,
+            "dependsOn": ["^build"],
+            "inputs": ["production", "^production"]
+        },
         "@nx/vite:build": {
             "cache": true,
             "dependsOn": ["^build"],
